Update how labels are drawn on the canvas
This is an attempt to push more plotting parameters to the config
file rather than hard-coding it in the plotting script. Specifically
you can now specify the legend location in the config file rather
than just a having a 'do legend' switch and you can specify the
labels you want drawn on the canvas.