Convert unicode keys and values from config file to strings
This fix addresses an issue with the TOML parser in Python 2 where
keys and values are returned in unicode format rather than standard
strings. This is necessary when passing strings to ROOT classes because
these delicate flowers get confused when they see anything that doesn't
resemble an array of raw ascii characters.